+++ ITZRealState/Models/Listing.cs
@@ -7,14 +7,14 @@ namespace ITZRealState.Models
     public class Listing
     {
         [Key]
-        public long IdListing { get; set; }
+        public int IdListing { get; set; }
         public string street { get; set; }
         public string number { get; set; }
         public string colony { get; set; }
         public double lotSize { get; set; }
         public double constSize { get; set; }
         public double price { get; set; }
-        public long IdUser { get; set;}
+        public int IdUser { get; set;}
         public bool sold { get; set; }
         public int zipcode { get; set; }
     }
